backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
alembic==1.13.3
celery[redis]==5.4.0
redis==5.1.1
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
python-multipart==0.0.12
httpx==0.27.2
groq==0.11.0
openai==1.51.0
Pillow==10.4.0
numpy==1.26.4
reportlab==4.2.5
cairosvg==2.7.1
pytesseract==0.3.13
cloudinary==1.41.0
razorpay==1.4.1
python-dotenv==1.0.1
aiofiles==24.1.0
pytest==8.3.3
pytest-asyncio==0.24.0
